#1239 update gocean and dynamo tests
